+++ src/orquestra/sdk/_base/cli/_dorq/_arg_resolvers.py
@@ -176,7 +176,9 @@ class SpacesResolver:
             return workspace_id
 
         workspaces = self._spaces_repo.list_workspaces(config)
-        labels = self._presenter.workspaces_list_to_prompt(workspaces)
+        labels, workspaces = self._presenter.workspaces_list_to_prompt(
+            workspaces, config
+        )
         prompt_choices = [(label, ws) for label, ws in zip(labels, workspaces)]
         selected_id = self._prompter.choice(prompt_choices, message="Workspace")
 
@@ -203,7 +205,7 @@ class SpacesResolver:
             return project_id
 
         projects = self._spaces_repo.list_projects(config, workspace_id)
-        labels = self._presenter.project_list_to_prompt(projects)
+        labels, projects = self._presenter.project_list_to_prompt(projects, config)
         if optional:
             projects.append(None)
             labels.append("All")
@@ -471,9 +473,12 @@ class WFRunFilterResolver:
             # The user has passed in one or more state arguments, iterate through them
             # and check that they're valid states.
             for state in states:
-                try:
-                    _ = State(state)
-                except ValueError:
+                s = State(state)
+
+                # If the state the user passed is not the same as we parsed, then this
+                # is an invalid state.
+                # This allows users to still ask for workflows in UNKNOWN states.
+                if state != s.value:
                     _invalid_states.append(state)
                 else:
                     _selected_states.append(state)