+++ tests/sdk/v2/data/sample_project/workflow_defs_no_raise.py
@@ -6,12 +6,12 @@ import helpers  # type: ignore
 import orquestra.sdk as sdk
 
 
-@sdk.task()
+@sdk.task
 def capitalize_task(text):
     return helpers.capitalize_helper(text)
 
 
-@sdk.workflow(default_source_import=sdk.InlineImport())
+@sdk.workflow
 def sample_workflow():
     return [capitalize_task("sample")]